Compound Protocol v2.2 Release Candidate
Note: **this is a release candidate version and is not yet fully completed audited. This code is liable for change. For concerns about security, message security@compound.finance**

This patch modifies the Compound Protocol Comptroller for a Timelock administrator which will add a delay to allow the community to audit and respond to changes before they are made. Additionally, in the event of an in-the-wild concern, we add a pause gaurdian that has the ability to pause borrows, withdraws or liqudations if any funds appear to be put at risk. We also change the behavior to allow a user to enter a market simply by invoking the borrow function.

Finally, we make a variety of changes in response to the OpenZeppelin audits for minor code quality concerns.