Add Compound Lens (#29)
* Add Compound Lens

The Compound Lens is a view contract Compound smart contracts that provides the data dApps need to stay hydrated. Specifically, we wrap the main data that our dApp is asking for. There's probably more bundling we could do, but this already reduces the raw number of calls by 98% (from about 10/s to about 0.2/s).

* Add CompoundLens deployments