Post-audit fixes, more scenarios, and auxiliary improvements fixes (#63)
* Compare boolean values numerically to 1, not string 0x01

* Remove the short-circuits for market NEVER initialized (#64)

These were intended to be short-circuits if not a COMP market, but in fact only take effect if *never* a COMP market. Removing the short-circuit avoids this confusion, and enables us to poke all borrowers once, instead of after every time a market becomes a COMP market for the first time.

* add checkIsComped, add isComped to markets storageAt getter (#65)

* Ensure updating COMP indices only takes 1 SSTORE (#66)

* Set claim threshold; scen code for forks (#69)

* Add events for changing comp rate and adding/dropping comped markets (#70)

* Rename dripper -> reservoir; add fauceteer (#72)

* Rename dripper -> reservoir

* Add Fauceteer and Add Timelock Harness Features (#71)

This patch adds the Fauceteer, which is a simple contract that can hold Erc20 tokens and will give away a small percentage every time you poke it. The idea is that it replaces the need for "FaucetToken" as we now just make Fauceteer own some amount of tokens and you can poke _it_ if you need test tokens. This gives us the ability to use real Dai, etc, on testnets *and* have a faucet available to users. As the Fauceteer only gives away 0.01% of what it has, it will effectively never run out, but instead just give away less and less each time you poke it. The raw number of tokens isn't that important for test-nets, so this is acceptable.

We also add some functions to the TimelockTest harness. We allow it to change admin without going through timelocking, and we add the ability for it to call `_acceptAdmin` without timelocking. This significantly makes the flow of setting up a test-net easier.

Co-authored-by: Max Wolff <maxcwolff@gmail.com>
Co-authored-by: Geoff Hayes <hayesgm@gmail.com>